Conversational RAG/requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
pypdf>=3.0.0
openai>=1.0.0
tiktoken>=0.5.0